requirements.txt
# Django Email API Requirements

# Core Django and REST Framework
django>=4.2.0,<5.0
djangorestframework>=3.14.0

# API Documentation
drf-yasg>=1.21.7

# Environment variables
python-dotenv>=1.0.0

# CORS support
django-cors-headers>=4.0.0

# Production server
gunicorn>=21.2.0

# Static files serving
whitenoise>=6.5.0

# Optional: Background tasks (uncomment if needed)
# celery>=5.3.0
# redis>=4.6.0

# Optional: File handling (uncomment if needed)
# pandas>=1.5.0           # For CSV handling
# openpyxl>=3.0.0         # For Excel support

# Note: This project uses SQLite by default (no additional database drivers needed)
# For production with PostgreSQL or MySQL, uncomment and install:
# psycopg2-binary>=2.9.0  # PostgreSQL
# mysqlclient>=2.2.0      # MySQL
